implement settable clean up option in gamma.proc

+++ process_ifg.bash
@@ -158,8 +158,10 @@ INT()
 	:
     fi
 
-    ## clean up unnecessary files
-    rm -f $ifg_offs $ifg_ccp $ifg_coffs $ifg_coffsets
+    if [ $clean_up == yes ]; then
+        ## clean up unnecessary files
+        rm -f $ifg_offs $ifg_ccp $ifg_coffs $ifg_coffsets
+    fi
 
     ## Create differential interferogram parameter file
     GM create_diff_par $ifg_off - $ifg_diff_par 0 0
@@ -278,8 +280,10 @@ FLAT()
         :
     fi
 
-    ## clean up unnecessary files
-    rm -f $ifg_base_temp $ifg_base_res $ifg_base_init $ifg_flat_temp $ifg_sim_unw0 $ifg_flat0
+    if [ $clean_up == yes ]; then
+        ## clean up unnecessary files
+        rm -f $ifg_base_temp $ifg_base_res $ifg_base_init $ifg_flat_temp $ifg_sim_unw0 $ifg_flat0
+    fi
 
 
     #######################################
@@ -342,11 +346,13 @@ FLAT()
         ## subtract simulated phase ('ifg_flat1' was originally 'ifg', but this file is no longer created)
 	GM sub_phase $ifg_flat1 $ifg_sim_unw $ifg_diff_par $ifg_flat 1 0
 
-        ## clean up unnecessary files
-        rm -f $ifg_flat1 $ifg_flat"1.unw" $ifg_sim_unw1 $ifg_flat1.unw 
-        rm -f $ifg_flat_cc0 $ifg_flat_cc0_mask 
-        rm -f $ifg_flat10.unw $ifg_off10 $ifg_flat10 $ifg_flat_cc10 $ifg_flat_cc10_mask 
-        rm -f $ifg_gcp $ifg_gcp_ph
+        if [ $clean_up == yes ]; then
+            ## clean up unnecessary files
+            rm -f $ifg_flat1 $ifg_flat"1.unw" $ifg_sim_unw1 $ifg_flat1.unw 
+            rm -f $ifg_flat_cc0 $ifg_flat_cc0_mask 
+            rm -f $ifg_flat10.unw $ifg_off10 $ifg_flat10 $ifg_flat_cc10 $ifg_flat_cc10_mask 
+            rm -f $ifg_gcp $ifg_gcp_ph
+        fi
     fi
 
     ## Calculate final flattened interferogram with common band filtering (diff ifg generation from co-registered SLCs and a simulated interferogram)
@@ -418,6 +424,12 @@ UNW()
 
         ## Use model to unwrap filtered interferogram
 	GM unw_model $ifg_filt $ifg_unw_model temp $ifg_width $ifg_refrg $ifg_refaz 0.0
+
+        if [ $clean_up == yes ]; then
+            ## clean up unnecessary files
+            rm -f $ifg_unw_thin $ifg_unw_model
+        fi
+
     else
         ## Unwrap the full interferogram without masking
         GM mcf $ifg_filt - - temp $ifg_width 1 - - - - $ifg_patch_r $ifg_patch_az - $ifg_refrg $ifg_refaz 1
